Refactor of downloading to try multiple pages of results. Fixes #10
This is a rather major rewrite of the logic around downloading files. Notably, I'm using the ES7 await/async syntax to make the logic around downloading a lot cleaner.  I've switched to axios instead of request (or request-promise-native) for better promises handling, but this has come at the expense of the progressbar, which is gone for now (#11). And in general, everything is just more cleanly using promises (e.g., we're now using mz for async file operations).

This does mean that the minimum required version of node is now 7.6.0, which I've added to the package.json.